+++ src/main/types.ts
@@ -166,8 +166,9 @@ type FileInfo = {
 };
 
 type VideoQueueItem = {
-  source: string;
-  suffix: string;
+  name: string; // Can an OBS timestamp if recording or more complicated if clipping.
+  source: string; // Can be either a path or a URL.
+  suffix: string; // Typically details of the recording, but can also be a "clipped at ..." description.
   offset: number;
   duration: number;
   clip: boolean;