+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 ### Fixed
 - Validate log path shows false by default when it's meant to show true.
+- Avoid shipping a duplicate copy of ffmpeg which reduces app size by about 60MB.
 
 ## [7.5.1] - 2026-01-21
 ### Fixed