+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 ### Added
 - [Issue 695](https://github.com/aza547/wow-recorder/issues/695) - Warn users when Advanced Combat Logging is disabled in WoW settings.
+- [Issue 797](https://github.com/aza547/wow-recorder/issues/797) - Add the new devourer spec.
+
 ### Fixed
 - Validate log path shows false by default when it's meant to show true.
 - Avoid shipping a duplicate copy of ffmpeg which reduces app size by about 60MB.