+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Changed
 ### Added
+### Fixed
+
+## [7.5.2] - 2026-02-26
+### Added
 - [Issue 797](https://github.com/aza547/wow-recorder/issues/797) - Add the new devourer spec.
 
 ### Fixed